45: Move types to component file
